+++ frontend/src/components/public/ProfilePicture.tsx
